refactor SymLinks cleaning in separate class
